Consolidate pose I/O in PoseIO
