+++ docs/pipelines.md
@@ -172,7 +172,7 @@ graph TD
     B -->|no| B1[Fail loudly<br/>list the unmatched file names]
     B -->|yes| C[Snapshot poses into Scene::priorPoses]
     C --> D{frames.json imported as AUTO?}
-    D -->|yes| E[DetectFramesConvention<br/>ImportFramesJSON.cpp]
+    D -->|yes| E[DetectFramesConvention<br/>PoseIO.cpp]
     E --> E1{Conclusive?}
     E1 -->|no| E2[Fail: ask for an explicit convention]
     E1 -->|flip needed| E3[FlipFramesConvention<br/>+ re-snapshot priorPoses]
@@ -188,7 +188,7 @@ graph TD
     L --> M[Shared tail of Scene::Reconstruct]
 ```
 
-- **Pose import** happens back in Step 1: `ImportConfig::importPosesFile` is dispatched on extension, `.csv` to `ImportPosesCSV()` and `.json` to `ImportFramesJSON()` (`libs/SFM/ImportFramesJSON.h`), before the camera de-duplication so identical per-frame intrinsics collapse into one shared `Camera`.
+- **Pose import** happens back in Step 1: `ImportConfig::importPosesFile` is dispatched on extension, `.csv` to `ImportPosesCSV()` and `.json` to `ImportFramesJSON()` (`libs/SFM/PoseIO.h`), before the camera de-duplication so identical per-frame intrinsics collapse into one shared `Camera`.
 - **Permissive first triangulation** (4× `maxReprojError`): the imported poses are approximate and the intrinsics may still be EXIF-derived, so the strict threshold would reject correct tracks before BA can fix the geometry.
 - **Clustering never runs** on this path; `Scene::priorPoses` is transient (not serialized) but is preserved by regular scene copies and moves.
 - **The shared tail still runs**, including the final `Resection::RegisterImages()` for images absent from the poses file, and closes with `AlignToPriorPoses()` in place of `AlignToGPS()`. Failure to estimate that final similarity fails the reconstruction rather than leaving successful output in an arbitrary bundle-adjustment gauge.
@@ -909,7 +909,7 @@ dMin, dMax: float           — depth range from SFM sparse points
 | `libs/SFM/GlobalAlignment.cpp` | Multi-scene merge |
 | `libs/SFM/SceneCluster.cpp` | Scene clustering |
 | `libs/SFM/ViewGraphCalibrator.cpp` | Focal length estimation |
-| `libs/SFM/ImportFramesJSON.cpp` | frames.json pose import + camera-axes convention detection |
+| `libs/SFM/PoseIO.cpp` | CSV/frames.json pose I/O + camera-axes convention detection |
 | `libs/SFM/KeyframeExtractor.cpp` | Video keyframe selection |
 | `libs/MVS/SceneDensify.cpp` | Dense depth estimation |
 | `libs/MVS/DepthMap.cpp` | CPU PatchMatch |
